+++ mlxtend/frequent_patterns/__init__.py
@@ -5,6 +5,7 @@
 # License: BSD 3 clause
 
 from .apriori import apriori
+from .fpgrowth import fpgrowth
 from .association_rules import association_rules
 
-__all__ = ["apriori", "association_rules"]
+__all__ = ["apriori", "association_rules", "fpgrowth"]